Add common function for user agent
