+++ ui/src/index.ts
@@ -21,3 +21,4 @@ export * from './wizard/ResourceGroupCreateStep';
 export * from './wizard/ResourceGroupListStep';
 export * from './wizard/StorageAccountListStep';
 export { registerUIExtensionVariables } from './extensionVariables';
+export * from './addExtensionUserAgentInfo';
