+++ ui/src/reportAnIssue.ts
@@ -6,24 +6,13 @@
 // tslint:disable-next-line:no-require-imports
 import opn = require("opn");
 import { IParsedError } from '../index';
-import { ext } from "./extensionVariables";
+import { getPackageInfo } from "./getPackageInfo";
 
 /**
  * Used to open the browser to the "New Issue" page on GitHub with relevant context pre-filled in the issue body
  */
 export function reportAnIssue(actionId: string, parsedError: IParsedError): void {
-    let packageJson: IPackageJson | undefined;
-    try {
-        // tslint:disable-next-line:non-literal-require
-        packageJson = <IPackageJson>require(ext.context.asAbsolutePath('package.json'));
-    } catch (error) {
-        // ignore errors
-    }
-
-    // tslint:disable-next-line:strict-boolean-expressions
-    const extensionName: string = (packageJson && packageJson.name) || 'vscode-azuretools';
-    // tslint:disable-next-line:strict-boolean-expressions
-    const extensionVersion: string = (packageJson && packageJson.version) || 'Unknown';
+    const [extensionName, extensionVersion]: [string, string] = getPackageInfo();
 
     const body: string = `
 Repro steps:
@@ -38,9 +27,4 @@ OS: ${process.platform}
 `;
     // tslint:disable-next-line:no-floating-promises
     opn(`https://github.com/Microsoft/${extensionName}/issues/new?body=${encodeURIComponent(body)}`);
-}
-
-interface IPackageJson {
-    version?: string;
-    name?: string;
 }
